Module dumbtls::bytes

Utilities for manipulating raw bytes within memory, and converting between these and human-readable encodings.

Structs

XorLengthError
Error returned if the user tries to xor two byte sequences of different lengths

Traits

SequenceXor
Trait for sequences of bytes that can be xor'd with other sequences of bytes.

Type Definitions

Bytes
The Bytes type represents an arbitrarily long sequence of raw bytes in memory.