Type Definition dumbtls::bytes::Bytes

type Bytes = Vec<u8>;

The Bytes type represents an arbitrarily long sequence of raw bytes in memory.

This type is a wrapper around a Vector containg 8-bit unsigned integers. This is used throughout DumbTLS to represent raw bytes, and to perform operations on these. While the actual differences to simply using a Vec are minimal, using this type semantically signfies that the integers stored within are a representation of memory, and may not carry any meaning in their decimal representation.

This type provides a number of methods for converting between bytes and human-readable encodings, such as hex and base64, and should be used when encoding/decoding user input & output.

Examples

Constructing Bytes from a hex string, and outputting the result as base64:

let my_bytes = Bytes::from_hex("cafebabe");
println!("Base64: {}", my_bytes.to_base64()); // Outputs: "Base64: yv66vg=="