Module dumbtls::ciphers::aes

Implementation of AES-128.

AES was developed as the result of a competition by NIST to produce a replacement to DES. Originally published by Vincent Rijmen and Joan Daemen in 1999 as Rijndael, NIST ratified it as being the Advanced Encryption Standard in 2001. AES is a block cipher, and uses a block size of 128 bits (= 16 bytes). It may use a key size of 128, 192, or 256 bits (= 16, 24, or 32 bytes), corresponding to to the designations AES-128, AES-192, and AES-256 respectively. AES uses either 10, 12, or 14 bits, corresponding to the increasing key sizes. AES is regarded as suitable for the majority of modern cryptographic applications where a symmetric block cipher is required, and finds widespread usage in TLS, SSH, and many other secure protocols. This is in part due to the fact that modern x86 processors include specific instructions for encryption/decryption with AES, which make it one of the most perfomant algorithms available.

Due to time constraints, I was unfortunately only able to implement AES-128, however I plan to implement AES-192 and AES-256 in the future, and the software is designed to be fairly easily extensible.

The design of AES relies heavily on Galois Field arithmetic: Briefly, a field is a (mathematical) set of elements, on which addition, subtraction, multiplication, and inversion are all defined and closed (the result of the operation is another element within the field). A Galois Field is a field which only has a finite number of elements. As such, the field operations must often be done modulo some element of the field, for closure to be retained. For prime fields (Galois fields where the number of elements is prime) this is simple: One just performs the operation modulo the size of the field. For extension fields (Galois fields where the number of elements is a prime to a certain power) this is not so simple. One must treat each element as a polynomial, and take the operations modulo some irreducible polynomial. AES works on the prime extension field $GF(2^8)$, and defines a single irreducible polynomial (the AES polynomial) which operators are performed modulo: $p(x) = x^8 + x^4 + x^3 + x + 1$. Christof Paar's Lecture on Galois Fields explains this in more depth.

Unlike DES, AES does not use a Feistel network design. Instead, each layer of encryption is actually inverted, and applied in reverse, in order to decrypt. When encrypting text with AES, the message is first Xor'd with the original key. Then, 10 rounds of encryption are applied, with the following steps being conducted in each round:

Byte Substitution: The Byte Susbstitution layer applies an S-Box to each byte of the input. This provides confusion, and introduces non-linearity to AES. Unlike with DES, the design of the AES S-Box is actually public, and is based in very specific algebra. The S-Box works as follows: First, the input is inverted in the Galois field $GF(2^8)$, using the AES polynomial. This inversion is defined for all elements except the zero byte, so this is instead simply mapped to itself. Using Galois field inversion provides a high degree of non-linearity, however, it has a strong algebraic structure, which could be used in cryptanalysis. Because of this, an affine mapping is then applied to the output of the inversion, which prevents this structure from being used. In the majority of implementations of AES, since one byte may only have 256 possible values, all possible mappings are pre-calculated and hardcoded, rather than calculating substitutions on the fly.

Diffusion: Two layers of diffusion are used within AES:

Shift Rows: The order of the bytes is altered: The message is split into 4-byte words, then bytes are moved between words such that each final word contains one byte from each original word.

Mix Columns: Each word is multiplied by a matrix of polynomials in $GF(2^8)$. In practise, this is implemented via a set of look-up tables.

Key Addition: A round subkey is derived from the original key, and the message is Xor'd with it.

To decrypt AES, the inverse of each operation is applied in reverse order.

Structs

AES
The AES Cipher.

Enums

AESError
Error encountered during AES encryption/decryption.

AESKey
Contains an AES key.

AESKeysize
Represents possible keysizes for use with AES.

Functions

key_to_len
Converts a key to the length it should be, in bytes.

keysize_to_len
Converts a keysize to the length a key of that keysize should be, in bytes.