Module dumbtls::ciphers::block

Modes of operation for symmetric block ciphers.

As discussed in the documentation for this module, block ciphers work using fixed-length blocks of input. However, we often wish to work with arbitrarily large (or small) input texts. For this, it may seem that block ciphers are not applicable: If an input is too long to fit in a block, it cannot be encrypted by a block cipher. It is for this reason that we implement generic "modes of operation" for block ciphers, which allow them to be used to encrypt and decrypt inputs of sizes greater than a single block.

These modes of operation are simply techniques for splitting the plaintext, and then applying the cipher multiple times to ensure the entire plaintext is encrypted. However, the choice of block cipher mode has important security implications, which must be considered when using them.

Structs

CBCMode
Cipher Block Chaining mode for block ciphers.

ECBMode
Electronic CodeBook mode for block ciphers.

Traits

BlockCipher
Trait for a generic block cipher.