Trait dumbtls::ciphers::block::BlockCipher

pub trait BlockCipher {
    type Error: Sized;
    fn get_block_size(&mut self, msg: &Bytes) -> usize;

    fn encrypt_block(&mut self, plaintext: Bytes) -> Result<Bytes, Self::Error>;

    fn decrypt_block(&mut self, ciphertext: Bytes) -> Result<Bytes, Self::Error>;
}

Trait for a generic block cipher.

This trait is to be implemented by structs representing a block cipher algorithm: They must be able to encrypt a single block of data, and decrypt a single block of data. They must also be able to provide their block size.

It is structs that implement this trait that are operated on by the block cipher modes of operation in the block module.

Associated Types

type Error: Sized

Error type returned when this cipher encounters an error.

Loading content...
Required methods

fn get_block_size(&mut self, msg: &Bytes) -> usize

Retrieve the block size of the cipher, in bytes.

fn encrypt_block(&mut self, plaintext: Bytes) -> Result<Bytes, Self::Error>

Encrypt the given plaintext.

This function should return the Ok option, containing the bytes representing the ciphertext if and only if the encryption was successful, otherwise the Err variant should be returned.

fn decrypt_block(&mut self, ciphertext: Bytes) -> Result<Bytes, Self::Error>

Decrypt the given plaintext.

This function should return the Ok option, containing the bytes representing the plaintext if and only if the decryption was successful, otherwise the Err variant should be returned.

Loading content...