Module dumbtls::ciphers::des

Implementation of DES.

DES is a block cipher based around a Feistel Network, and as such, it is implemented in DumbTLS using the FeistelCipher and FeistelNetwork structures.

DES was originally developed by IBM in the 1970s, with input from the NSA, and was published for public use. Being one of the first modern ciphers to be made publicly available, DES was widely used throughout the late 20th century. The cipher uses a key size of 64 bits (= 8 bytes), although 8 of those bits are not used during encryption, so it should really be thought of as using a 56-bit key. DES operates on blocks of size 64 bits. This cipher is now known to be insecure due to its short keysize, and should not be used for cryptographic purposes.

As DES is based on a Feistel Network, in each round, the right hand side is passed through a round function, to derive an intermediate value with which the left hand side is Xor'd. The round function combines the RHS with a round key. The process for doing so is as follows:

Expansion: The 32-bit right hand side is expanded from 8 4-bit blocks (= 32 bits) to 8 6-bit blocks (= 48 bits). This is done to increase diffusion: 16 of the 32 bits appear twice in the output of the expansion, but an input bit never occurs twice in the same 6-bit output block. This bit mapping is done according to the DES expansion permutation table, E.

XOR: A 48-bit subkey for this round is derived from the original key, and the output of the expansion is Xor'd with it.

Substitution: The 48-bit result of the XOR is split into 8 6-bit blocks, each of which is passed through a Subsitution Box (or S-Box). The S-Boxes provide confusion for the cipher. They also provide non-linearity, which defends DES against known plaintext attacks. Each S-Box maps 6 bits to 4 bits, so the output of the substitution is 32 bits long.

Permutation: Permutation is used to ensure that the outputs of the S-Boxes will pass through different S-Boxes in the next round, increasing diffusion. It is a 1:1 mapping of bits, so its output is 32 bits long.

The NSA are known to have heavily influenced the design of the S-Boxes for the substitution round, which was originally a major concern for many, due to the possibility of there being a backdoor. However, it was later revealed that the NSA's input was intended to strengthen DES against the (at the time) non-public differential cryptanalysis attack.

Structs

DESBlock
DES internal block ciper.

Type Definitions

DES
DES cipher