Struct dumbtls::ciphers::des::DESBlock

pub struct DESBlock {
    pub key: Bytes,
    // some fields omitted
}

DES internal block ciper.

Used within a FeistelNetwork instance. Please don't use this directly: Use DES instead.

Fields
key: Bytes
Implementations

impl DESBlock

pub fn get_bit(&self, byte: u8, bit_pos: u8) -> u8

Get a given bit's value (0 or 1) from an integer.

pub fn map_bits(&self, unmapped_bytes: &mut Bytes, map: &[usize])

Maps the bits of a collection of Bytes to new locations.

Takes an instance of Bytes and a slice containing the indices of bits within the first collection, in the positions where they should be placed. Mutates the first to move bits to where they should go.

pub fn initial_permutation(&self, plaintext: &mut Bytes)

Perform the initial permutation on plaintext, prior to performing the DES rounds.

pub fn final_permutation(&self, ciphertext: &mut Bytes)

Perform the final permutation on ciphertext, after having performed the DES rounds.

pub fn permuted_choice_1(&mut self)

Perform permuted choice 1, to initialise the key, before performing encryption/decryption.

pub fn rotate_key_left(&self, key: &Bytes) -> Bytes

Rotate the two halves of the key one place left.

Each half of the key must be rotated individually. Unfortunately, the break between the two halves falls in the middle of the byte, so we've got to do some pretty arcane stuff to get this working.

pub fn permuted_choice_2(&self, round_num: u32) -> Bytes

Perform permuted choice 2 to derive a round subkey.

pub fn half_block_expansion(&self, rhs: &mut Bytes)

Expand the right hand side, for use in the round function.

pub fn six_byte_split(&self, rhs: &Bytes) -> Bytes

Splits 6 8-bit bytes into 8 6-bit bytes.

pub fn apply_s_box(&self, byte: u8, s_box: usize) -> u8

Applies an S-box to convert a 6-bit byte to a 4-bit byte.

pub fn apply_p_box(&self, rhs: &mut Bytes)

Apply permutation at the end of the round function.