Module dumbtls::ciphers::feistel

Traits and structs for implementing ciphers built around Feistel Networks.

Feistel networks are a generic structure for implementing block ciphers with heavy confusion and diffusion, originally designed by Horst Feistel. For a cipher to implement confusion and diffusion, a common technique is to perform multiple "rounds": In each round, the key is used to encrypt the text somehow, then the text is shuffled round, with the next round depending on the previous round. With sufficient rounds, this technique can be very easily used to implement very strong confusion and diffusion, however, to reverse the encryption, it is necessary to invert each round, which may be mathematically impossible.

Feistel networks provide a simple solution to implementing multiple-round ciphers. In each round, the text to encrypt is split into two halves. The second half is run through a round function with the key as an input, and the result is then Xor'd with the first half. The output of the round is then created by concatenating the second half (entirely unaltered), with this Xor'd first half, so that the alteed first half is now placed second. In the next round, the same process is done, so the new second half is run through the round function, then Xor'd with the new first half (which was the second half in the previous round). The output of each round only encrypts half of the message. This means that a Feistel cipher is extremely easy to invert, given the key: You simply do the same operations in reverse.

The layout of a Feistel cipher is shown in the diagram below, with F denoting the round function.

      Encryption:          Decryption:
    L0           R0    RN+1          LN+1
    |             |      |             |
    |     Key     |      |     Key     |
    |      |      |      |      |      |
    v      v      |      v      v      |
   Xor <-- F <----+     Xor <-- F <----+
    |             |      |             |
    v             v      v             v
    R1           L1      LN           RN
    |             |      |             |
    |     Key     |      |     Key     |
    |      |      |      |      |      |
    |      v      v      |      v      v
    +----> F --> Xor     +----> F --> Xor
    |             |      |             |
    v             v      v             v
    L2           R2    RN-1          LN-1
         . . .                . . .
    |             |      |             |
    v             v      v             v
    LN           RN      R1           L1
    |             |      |             |
    |     Key     |      |     Key     |
    |      |      |      |      |      |
    v      v      |      v      v      |
   Xor <-- F <----+     Xor <-- F <----+
    |             |      |             |
    v             v      v             v
   RN+1         LN+1     L0           R0

The Feistel Network structure is made use of by a number of well known ciphers, including DES and RSAES-OAEP.

Structs

FeistelNetwork
Feistel Network implementation

Enums

FeistelCipherError
Errors encountered while encrypting/decrypting using a Feistel Network.

Traits

FeistelCipher
Trait for ciphers which work using a Feistel network.

Type Definitions

FeistelResult