Trait dumbtls::ciphers::feistel::FeistelCipher

pub trait FeistelCipher {
    fn get_num_rounds(&mut self, plaintext: &Bytes) -> u32;

    fn get_block_size(&mut self, plaintext: &Bytes) -> usize;

    fn derive_round_subkey(&mut self, round_num: u32) -> FeistelResult<Bytes>;

    fn round_function(
        &mut self, 
        rhs: &Bytes, 
        subkey: &Bytes, 
        round_num: u32
    ) -> FeistelResult<Bytes>;

    fn is_valid_plaintext_block(
        &mut self, 
        plaintext: &Bytes
    ) -> FeistelResult<()>;

    fn split_plaintext(&mut self, plaintext: &Bytes) -> (Bytes, Bytes);

    fn is_valid_ciphertext_block(
        &mut self, 
        ciphertext: &Bytes
    ) -> FeistelResult<()>;

    fn split_ciphertext(&mut self, ciphertext: &Bytes) -> (Bytes, Bytes);

    fn pre_encrypt_block(&mut self, _plaintext: &mut Bytes) { ... }

    fn post_encrypt_block(&mut self, _ciphertext: &mut Bytes) { ... }

    fn pre_decrypt_block(&mut self, ciphertext: &mut Bytes) { ... }

    fn post_decrypt_block(&mut self, plaintext: &mut Bytes) { ... }
}

Trait for ciphers which work using a Feistel network.

Types implementing this trait are intended for internal use within the FeistelNetwork struct, which runs the actual encryption/decryption using the primitives provided by the FeistelCipher instance.

Required methods

fn get_num_rounds(&mut self, plaintext: &Bytes) -> u32

Get the number of rounds to perform.

At least two rounds are necessary for the entire message to have been encrypted. The FeistelNetwork will run this number of rounds.

fn get_block_size(&mut self, plaintext: &Bytes) -> usize

Get the block size used by the cipher.

fn derive_round_subkey(&mut self, round_num: u32) -> FeistelResult<Bytes>

Derive the subkey for a given round.

This subkey will be provided to the round function.

fn round_function(
&mut self,
rhs: &Bytes,
subkey: &Bytes,
round_num: u32
) -> FeistelResult<Bytes>

Perform the round function, and return the data to be Xor'd against the current left hand side.

fn is_valid_plaintext_block(&mut self, plaintext: &Bytes) -> FeistelResult<()>

Retuns an error if the plaintext is invalid.

fn split_plaintext(&mut self, plaintext: &Bytes) -> (Bytes, Bytes)

Split the plaintext into left and right hand sides.

These don't necessarily have to be of equal length, so long as the round function can deal with that an output appropriate-length data.

fn is_valid_ciphertext_block(&mut self, ciphertext: &Bytes) -> FeistelResult<()>

Returns an error if the ciphertext is invalid.

fn split_ciphertext(&mut self, ciphertext: &Bytes) -> (Bytes, Bytes)

Split the ciphertext into left and right hand sides.

These don't necessarily have to be of equal length, so long as the round function can deal with that an output appropriate-length data.

Loading content...
Provided methods

fn pre_encrypt_block(&mut self, _plaintext: &mut Bytes)

Perfom any transformations to the plaintext that are required before encryption begins.

Don't raise any errors here, do that in the is_valid_plaintext_block method.

fn post_encrypt_block(&mut self, _ciphertext: &mut Bytes)

Perform any transformations to the ciphertext that are required after encryption ends.

fn pre_decrypt_block(&mut self, ciphertext: &mut Bytes)

Perfom any transformations to the ciphertext that are required before decryption begins.

Don't raise any errors here, do that in the is_valid_ciphertext_block method.

Default implementation just calls [self.pre_encrypt_block]

fn post_decrypt_block(&mut self, plaintext: &mut Bytes)

Perfom any transformations to the plaintext that are required after decryption ends.

Default implementation just calls [self.post_encrypt_block]

Loading content...