Module dumbtls::ciphers::oaep

Implementation of RSAES-OAEP (RSA with Optimal Asymmetric Encryption Padding).

This module implements RSAES-OAEP as outlined in PKCS #1 (version 2.2).

RSA, as initially described within the academic literature, is not suitable for cryptographic use: Encryption is entirely deterministic, so the same plaintext will always encrypt to the same ciphertext when the same key is used. Furthermore, RSA is highly mathematically structured, and has some vulnerabilities (e.g: via multiplicativity) due to this. It is for this reason that RSA must only be used with a secure padding scheme. One such scheme, which is recommended for modern use, is Optimal Asymmetric Encryption Padding (OAEP).

OAEP uses a Feistel network in order to introduce a random element to RSA encryption: a random seed is generated, with fixed length (normally equal to the output length of a hash function). The message to be encrypted is padded with zeroes, until the total length of the input (message + padding + seed) equals the length of the RSA modulus to be used. The message and padding are concatenated to form the left hand side of the input for the Feistel Network, and the seed forms the right hand side. A mask generating function is used to form the round function: This uses the selected hash function to generate an arbitrary sized mask from its input. First a mask is generated from the random seed, then Xor'd with the message & padding to form the left hand side of the output. Then, a mask is generated from this output, and Xor'd with the random seed to form the right hand side of the output.

To reverse OAEP, simply reverse the Feistel network, as normal.

Example usage

use dumbtls::bytes::Bytes;
use dumbtls::ciphers::block::BlockCipher;
use dumbtls::ciphers::oaep::OAEP;
use dumbtls::ciphers::rsa::{RSAKeysize, RSAKey};
use dumbtls::encoding::hex::{ToHex, FromHex};
use dumbtls::keygen::gen_key_rsa;
 
fn main() {
    let keypair = gen_key_rsa(RSAKeysize::Key1024Bit);
    println!("Key modulus: {}", keypair.public.get_modulus());
    println!("Public exponent: {}", keypair.public.get_exponent());
    println!("Private exponent: {}", keypair.private.get_exponent());
    let plaintext = Bytes::from_hex("cafebabe").unwrap();
    println!("Plaintext: {}", plaintext.to_hex());
    let pubkey = keypair.public.clone();
    let mut enc_cipher = OAEP::new(pubkey);
    let ciphertext = enc_cipher.encrypt_block(plaintext).unwrap();
    println!("Ciphertext: {}", ciphertext.to_hex());
    let mut dec_cipher = OAEP::new(keypair.private.clone());
    let plaintext = dec_cipher.decrypt_block(ciphertext).unwrap();
    println!("Decrypted ciphertext: {}", plaintext.to_hex());
    // Example output:
    // Key modulus: 92923642353856878160738108776523986341734... (Truncated)
    // Public exponent: 65537
    // Private exponent: 379613905318714845754840008795561112... (Truncated)
    // Plaintext: cafebabe
    // Ciphertext: 5bf31b4ed6a1f5e53684ef0d738d5595c89dcbe33b... (Truncated)
    // Decrypted ciphertext: cafebabe
}

Structs

MGF1

OAEPBlock

Traits

MaskGenerationFunction

Type Definitions

OAEP