Module dumbtls::ciphers::rsa

RSA encryption/decryption primitives.

This module implements the data conversion and cryptographic primitives for RSA outlined in PKCS #1. Textbook RSA, as implemented in this module, is not suitable for cryptographic use. See the OAEP module for an implementation of Optimal Asymmetric Encryption Padding (RSAES-OAEP) according to PKCS #1. OAEP is designed to make textbook RSA safe for use in real-world encryption.

RSA is one of the most important ciphers ever developed. Originally developed by Clifford Cocks of GCHQ (and kept secret), then later independently developed and publicly released by Rivest, Shamir, and Adleman; RSA is believed to be the first public key asymmetric cryptosystem to have ever been developed, and was certainly the first to be publicly released. While RSA is now less used, in favour of elliptic curve cryptography, it laid the groundwork for many cryptosystems we rely on every day in the modern age, and is still supported in many cryptographic systems.

RSA's security relies on the difficulty of finding the prime factors of large numbers. This is a well known problem in computational number theory: For certain semi-primes (the product of two prime numbers), where each of the two prime factors is sufficiently large, it is computationally infeasible to factor this number (no polynomial-time algorithm exists).

To understand RSA, it is necessary to have some background in number theory, specifically in modular arithmetic: Khan Academy's Modular Arithmetic Course is more than sufficient to understand all that is necessary for RSA. It is also necessary to be familiar with Euler's Totient Function, $\phi(n): \mathbb{N} \mapsto \mathbb{N}$. Euler's totient function simply returns the number of integers less than its input which are coprime to that input (i.e: their greatest common denominator with $n$ is 1). From Euler's theory, for $n, a \in \mathbb{N}: \gcd(a, n) = 1$, we have that:

$$ a^{\phi(n)} \equiv 1 \mod{n} $$

We can now begin to understand how the algorithm works. First, two large prime numbers, $p, q$ are chosen, and kept secret. We calculate their product, $n = pq$, which is the first half of the public key, called the modulus. It's okay for $n$ to be made public, since there's no efficient way to calculate $p$ or $q$ given $n$. We next calculate $\phi(n)$. This is computationally infeasible given only $n$, but we (helpfully!) also know $n$'s prime factors. For all prime numbers $a, \phi(a) = a - 1$, since all numbers less than a prime are coprime to it. Furthermore, for $d = \gcd(a, b), \phi(ab) = \phi(a) \cdot \phi(b) \cdot \frac{d}{\phi(d)}$. We know that $n = pq$, so (since the gcd of two primes is always 1):

$$ \begin{aligned} \phi(n) & = \phi(p) \cdot \phi(q) \cdot \frac{\gcd(p, q)}{\phi(\gcd(p, q))} \\ & = \phi(p) \cdot \phi(q) \\ & = (p - 1) \cdot (q - 1) \end{aligned} $$

It is important that this value is kept secret. We now choose the public exponent, $e$, an integer that constitutes the second half of the public key. The public exponent is chosen such that it is relatively prime to $\phi(n)$. In practice, we normally set the public exponent to a fixed value, and then choose primes such that $\gcd(e, \phi(n)) = 1$, rather than choosing the primes first.

Having selected $e$, we calculate its modular inverse: $d = e^{-1} \mod \phi(n)$, so $de \equiv 1 \mod \phi(n)$. This was why $e$ had to be coprime to $\phi(n)$: The modular inverse only exists if $e$ and $\phi(n)$ are coprime. $d$ is called the private exponent, and together with $n$, makes up the private key. We now have everything necessary to conduct RSA.

The actual RSA algorithm operates on integers, not bytes. The RSA standards define an algorithm for converting a byte string to an integer, which we will not cover here, but is implemeneted within this module. To encrypt a message $m \in \mathbb{N}$ with the public key, simply calculate $c = m^e \mod n$. (n.b: $m$ must be less than $n$, and $m^e$ must be greater than $n$). $c$ is the ciphertext, and may be released publicly. Due to the discrete logarithm problem, $m$ cannot be recovered from $c$, without the use of the private key. To decrypt, simply compute $m = c^d \mod n$.

Why does this work? Recall Euler's theorem: $a^{\phi(n)} \equiv 1 \mod{n}$. We know that $de \equiv 1 (\mod \phi(n))$, which can also be written as $de = k \cdot \phi(n) + 1$ for some $k \in \mathbb{N}$. So, $m^{de} = m^{k \cdot \phi(n) + 1}$ for some $k$. Now:

$$ \begin{aligned} m^{k \cdot \phi(n) + 1} & \equiv m \cdot m^{k \cdot \phi(n)} \mod n \\ & \equiv m \cdot (m^{\phi(n)})^k \mod n \\ & \equiv m \cdot 1^k \mod n \\ & \equiv m \mod n \end{aligned} $$

Throughout this description, we have ignored an important development in computational number theory which has occured since the development of RSA: quantum computation. Shor's Algorithm, invented in 1994, provides a polynomial-time algorithm for factorising integers. While quantum computers are not yet powerful enough to factorise numbers on the scale of those used by RSA, the possibility of RSA's key assumptions being undermined must be a serious consideration for anyone using RSA in the modern age.

Example usage (without padding)

See the OAEP module for an example using padding.

use dumbtls::bytes::Bytes;
use dumbtls::ciphers::{rsa, rsa::RSAKey};
use dumbtls::encoding::hex::{FromHex, ToHex};
use dumbtls::keygen::gen_key_rsa;

fn main() {
    let keypair = gen_key_rsa(rsa::RSAKeysize::Key512Bit);
    println!("Key modulus: {}", keypair.public.get_modulus());
    println!("Public exponent: {}", keypair.public.get_exponent());
    println!("Private exponent: {}", keypair.private.get_exponent());
    let plaintext = Bytes::from_hex("cafebabe").unwrap();
    println!("Plaintext: {}", plaintext.to_hex());
    let plaintext_int = rsa::bytes_to_integer(&plaintext);
    let ciphertext_int = rsa::encrypt_int(&keypair.public, plaintext_int).unwrap();
    let ciphertext = rsa::integer_to_bytes(ciphertext_int.clone(), 64).unwrap();
    println!("Ciphertext: {}", ciphertext.to_hex());
    let plaintext_int = rsa::decrypt_int(&keypair.private, ciphertext_int).unwrap();
    let plaintext = rsa::integer_to_bytes(plaintext_int.clone(), 4).unwrap();
    println!("Decrypted ciphertext: {}", plaintext.to_hex());
    // Example output:
    // Key modulus: 78445037715101259013909099418635822507710... (Truncated)
    // Public exponent: 65537
    // Private exponent: 506971569101898748116959870893742506... (Truncated)
    // Plaintext: cafebabe
    // Ciphertext: 0006b41d33536de82c6e30694c1af6553e1b831e80... (Truncated)
    // Decrypted ciphertext: cafebabe
}

Structs

RSAKeypair
Represents an RSA keypair (public & private).

RSAPrivateKey
Represents an RSA private key.

RSAPublicKey
Represents an RSA public key.

Enums

RSAError
Errors that can occur while attempting to perform RSA.

RSAKeysize
Represents the size of an RSA key.

Traits

RSAKey
Trait for types which contain RSA keys.

Functions

bytes_to_integer
Convert a sequence of octets to an integer.

decrypt_int
Decrypt an integer with the given RSA private key.

encrypt_int
Encrypt an integer with the given RSA public key.

integer_to_bytes
Convert an integer to a sequence of octets.

keysize_map
Maps RSA keysizes to the length of the modulus, in bytes.

sign_int
Sign an integer with the given RSA private key.

verify_sig_int
Verify an integer has been signed by a given RSA private key, using the corresponding public key.

Type Definitions

RSAResult