Module dumbtls::encoding::base64

Traits for encoding/decoding base64 data.

Modules

mappings
This module contains a number of common mappings used for base64-encoded data.

Structs

Config
Contains configuration parameters for base64 encoding

Enums

CharacterSet
Available encoding character sets

DecodeError
Represents an error encountered while decoding a base64 string to bytes.

Traits

FromBase64
Implemented for types which can be instantiated from base64-encoded data.

ToBase64
Implemented for types which contain raw bytes, and can output these raw bytes as base64-encoded data.