Module dumbtls::encoding::base64::mappings

This module contains a number of common mappings used for base64-encoded data.

Different implementations of base64-encoding use different sets of 64 characters to encode the data. You may wish to decode from a source that uses one of these different encodings, rather than the standard mapping, or encode for use in such an encoding. Each member of this module is a Config instance, and can be passed to FromBase64::from_base64_config or ToBase64::to_base64_config to customise the mapping used.

These mappings are simply re-exported from the base64 library.

Constants

BCRYPT
Bcrypt character set

BINHEX
BinHex character set

CRYPT
As per crypt(3) requirements

IMAP_MUTF7
IMAP modified UTF-7 requirements

STANDARD
Standard character set with padding.

STANDARD_NO_PAD
Standard character set without padding.

URL_SAFE
URL-safe character set with padding

URL_SAFE_NO_PAD
URL-safe character set without padding