Trait dumbtls::encoding::hex::FromHex

pub trait FromHex {
    fn from_hex(src: &str) -> Result<Self, DecodeError>
    where
        Self: Sized;
}

Implemented for types which can be instantiated from hex-encoded data.

Required methods

fn from_hex(src: &str) -> Result<Self, DecodeError> where
Self: Sized,

Create a new instance of the type implementing this trait from the hex-encoded data src.

Decodes the hex-encoded string src into raw bytes, and returns a Result containing a new instance of the type if the decoding was successful, or a DecodeError otherwise.

Errors

This function will return a DecodeError when the input src is not valid hex-encoded data: This may occur if the string passed to the function has non-even length (as valid hex data uses two digits to represent one byte, corresponding to the OddLength variant), or if there is aninvalid character in the string (i.e: Not 0-9, A-F, or a-f, corresponding to the InvalidHexCharacter variant).

Loading content...