Trait dumbtls::encoding::hex::ToHex

pub trait ToHex {
    fn to_hex(&self) -> String;

    fn to_hex_upper(&self) -> String;
}

Implemented for types which contain raw bytes, and can output these raw bytes as hex-encoded data.

Required methods

fn to_hex(&self) -> String

Encode the data contained within the type as lower-case hex.

This method should not cause any errors.

fn to_hex_upper(&self) -> String

Encode the data contained within the type as upper-case hex.

This method should not cause any errors.

Loading content...