Module dumbtls::hashes

Hash functions.

The content of this module is, unfortunately, not currently implemented in DumbTLS, and we instead rely on the external sha-1 and sha2 crates to provide the algorithms. In the future, I plan to actually implement these hash functions (and more!) but for the time being, dependencies are used.

Modules

sha1

sha2

Traits

HashFunction
Trait for a generic hash function.