Module dumbtls::keygen

Functions for generating keys for use in ciphers.

Different ciphers have different requirements as to the keys they use for encryption/decryption. This module is intended to provide utilities for generating (cryptographically secure) pseudo-random keys for use with the ciphers implemented in DumbTLS.

The random-number generation used in this module is implemented within the rand crate, which uses the Operating System's entropy pool to seed a CSPRNG (Cryptographically Secure Pseudo-Random Number Generator).

Functions

gen_key_aes

gen_key_des

gen_key_rsa