Module dumbtls::padding

Utilities for adding/removing padding from bytes.

Many modern ciphers are categorised as "block ciphers": they encrypt inputs of a specific, fixed block size, rather than encrypting any sized input. This often allows for easier implementation of important cryptographic principals, such as diffusion, since the entire block can be operated upon at once. However, this does raise an issue for end users: We often wish to encrypt content that is not an exact multiple of the block size in length.

Padding is used to resolve this issue. We take the input, however long it is, and add extra characters until its length is a multiple of the input size. For a padding scheme to be effective, it must be easily removable: Simply adding random bytes to the end of some input could definitely work to pad it, but when the time comes to remove the padding, how can you tell where the input ends and the padding begins? Padding schemes will often include the length of the padding as part of the padding itself, or use characters which cannot occur in a valid input, if that input is of a certain type.

Furthermore, some ciphers, such as RSA, have deeper issues relating to padding. While RSA does not have a set block-size, "textbook RSA", in which input is simply encrypted without alteration, is actually deeply flawed in terms of security. It is necessary to alter the input in order for it to be secure, which is done via a specific padding scheme.

Modules

null
Functions for adding/removing null-byte padding.

pkcs7
Functions for adding/removing padding according to the PKCS #7 standard

Structs

NULL

PKCS7

Traits

PaddingScheme