Module dumbtls::padding::null

Functions for adding/removing null-byte padding.

In some programming languages, strings are terminated by a null-byte (i.e: 0x00, the byte with all bits set to false). Therefore, if one is padding string data for use with a block cipher, it is often convenient to pad the data with repeated null bytes, up until it has length equal to a multiple of the block size. To remove the padding, simply remove all null bytes from the end of the final block.

Functions

pad
Pad the given content using null bytes.

unpad
Remove null padding from the given content.