Module dumbtls::padding::pkcs7

Functions for adding/removing padding according to the PKCS #7 standard

The Public Key Cryptography Standards (PKCS) are a set of standards, originally outlined by RSA Security, used to define how different systems should implement cryptographic protocols, so that interoperability is possible. PKCS #7, defined in RFC 2315, outlines syntax for transmitting encrypted data. One detail included in this document is a padding scheme for use with block ciphers, of any block-size < 32 bytes. This padding scheme, often referred to simply as "PKCS #7 padding" has become commonplace when using block ciphers, in part due to its simplicity.

The algorithm for padding data is as follows: Suppose that a block cipher has block-size $k \in \mathbb{N}$ bytes. To perform PKCS #7 padding on an input of length $l \in \mathbb{N}$ bytes , add $k - (l\mod k)$ binary octets to the end of the input, with each octet having value $k - (l\mod k)$. It should be noted that, since $nk\mod k = 0 \enspace \forall \enspace n \in \mathbb{N}$, an extra block, filled with octets of value $k$, is added to the input. This behaviour (an extra block being added) does not happen for any other input sizes.

To give an example of this padding in use, consider some block cipher with block size 8 bytes ($= k$). We wish to encrypt the (hex-encoded) data 0xabcdef0011. The length of this data is 5 bytes ($= l$), so $l\mod k = 5,$ and $k - (l\mod k) = 3$. The hex-encoding of 3 is 0x03, so we need to add $k - (l\mod k) = 3$ octets of data, each with the value 0x03. Therefore, the input to be sent to the block cipher for encryption is: 0xabcdef0011030303.

To remove the padding from data, so that it may be used, we simply need to consider the value of the last octet, then remove that many bits from the end of the data.

PKCS #7 padding may also be referred to as Cryptographic Message Syntax (CMS) padding, or PKCS #5 padding.

Structs

InvalidPadding
Represents an error encountered when trying to remove padding.

Functions

pad
Pad the given content using PKCS #7.

unpad
Remnove PKCS #7 padding from the given content.